requirements.txt
streamlit==1.32.0
yfinance==0.2.28
pandas-ta==0.3.14b0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
